rename project to match mod name
