+++ StardewValleyEsp/Detectors/EntityList.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+using Microsoft.Xna.Framework;
+
+namespace StardewValleyEsp.Detectors
+{
+    class EntityList : List<KeyValuePair<Vector2, object>>
+    {
+    }
+}
