+++ StardewValleyEsp/Detectors/ObjectDetector.cs
@@ -1,9 +1,9 @@
 ﻿using System.Collections.Generic;
 using Microsoft.Xna.Framework;
-using sdv_helper.Config;
 using StardewValley;
+using StardewValleyEsp.Config;
 
-namespace sdv_helper.Detectors
+namespace StardewValleyEsp.Detectors
 {
     class ObjectDetector : IDetector
     {
