+++ StardewValleyEsp/Menu/ConfigMenu.cs
@@ -1,16 +1,16 @@
-﻿using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
-using Microsoft.Xna.Framework.Input;
-using sdv_helper.Config;
-using StardewModdingAPI;
-using StardewValley;
-using StardewValley.Menus;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
+using Microsoft.Xna.Framework.Input;
+using StardewModdingAPI;
+using StardewValley;
+using StardewValley.Menus;
+using StardewValleyEsp.Config;
 
-namespace sdv_helper.Menu
+namespace StardewValleyEsp.Menu
 {
     class ConfigMenu : IClickableMenu
     {
