+++ StardewValleyEsp/Menu/Tab.cs
@@ -3,7 +3,7 @@ using Microsoft.Xna.Framework.Graphics;
 using StardewValley;
 using StardewValley.Menus;
 
-namespace sdv_helper.Menu
+namespace StardewValleyEsp.Menu
 {
     class Tab : IClickableMenu
     {
