+++ StardewValleyEsp/StardewValleyEsp.csproj
@@ -1,8 +1,8 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <AssemblyName>sdv-helper</AssemblyName>
-    <RootNamespace>sdv_helper</RootNamespace>
+    <AssemblyName>StardewValleyEsp</AssemblyName>
+    <RootNamespace>StardewValleyEsp</RootNamespace>
     <Version>1.1.0</Version>
     <TargetFramework>net452</TargetFramework>
     <Platforms>x86</Platforms>
