+++ StardewValleyEsp/manifest.json
@@ -4,7 +4,7 @@
   "Version": "1.1.0",
   "Description": "Adds name tags to in-game entities for easier detection and finding.",
   "UniqueID": "reimu.sdv-helper",
-  "EntryDll": "sdv-helper.dll",
+  "EntryDll": "StardewValleyEsp.dll",
   "MinimumApiVersion": "2.11.2",
   "UpdateKeys": [
     "GitHub:rei2hu/stardewvalley-esp",
