--- sdv-helper/Detectors/EntityList.cs
@@ -1,9 +1,0 @@
-﻿using Microsoft.Xna.Framework;
-using System.Collections.Generic;
-
-namespace sdv_helper.Detectors
-{
-    class EntityList : List<KeyValuePair<Vector2, object>>
-    {
-    }
-}
